Extracted some home screen components out
